requirements-ml.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0

# Medical imaging
pydicom>=2.4.0

# Metrics and evaluation
scikit-image>=0.21.0

# Logging and monitoring
tensorboard>=2.14.0

# Configuration
pyyaml>=6.0

# Data handling
tqdm>=4.66.0
